backend/pyproject.toml
[project]
name = "davinci-design-backend"
version = "0.1.0"
description = "Backend for davinci-design web app"
authors = [
    {name = "Kye Burchard", email = "kyeb@mit.edu"},
]
dependencies = [
    "strawberry-graphql[fastapi]>=0.227.2",
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.29.0",
    "boto3>=1.34.93",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.30",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
